backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.9
pgvector==0.3.2

# ── Validation ────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Auth ──────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ── Azure ─────────────────────────────────────────────────────
azure-ai-formrecognizer==3.3.3
openai==1.51.0

# ── Rate limiting ─────────────────────────────────────────────
slowapi==0.1.9

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.2